edge-capital-partners-main/requirements.txt
pyyaml
python-dotenv
ib_insync
pyTelegramBotAPI
certifi
imapclient
aiohttp
pandas
ccxt